created ad_notifications.py for sending messages to emails
