+++ main/models.py
@@ -1,12 +1,12 @@
 from django.db import models
 from django.contrib.auth.models import User, Group
-from django.core.mail import send_mail
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 
 from pytils import translit
 
 from main.validator import validate_itn
+from main.ad_notifications import start_emailing
 
 
 class BaseModel(models.Model):
@@ -95,19 +95,4 @@ def create_user_profile(sender, instance, created, **kwargs):
         instance.groups.add(Group.objects.get(name="common users"))
 
 
-@receiver(post_save, sender=Ad)
-def send_new_ad_notification_email(sender, instance, created, **kwargs):
-    user_emails = [user.email for user in User.objects.exclude(subscription=None)]
-    company_email = "badwolfproduction.com"
-    if created:
-        name = instance.name
-        subject = "Новое объявление"
-        message = f"Кто-то продает {name}. Спеши посмотреть!"
-
-        send_mail(
-            subject,
-            message,
-            company_email,
-            user_emails,
-            fail_silently=False,
-        )
+start_emailing(Ad)
