+++ requirements.txt
@@ -4,7 +4,5 @@ django-constance==2.8.0
 pytils==0.3
 Pillow==8.4.0
 sorl-thumbnail==12.7.0
-django-allauth
-
-
-
+django-allauth==0.46.*
+apscheduler==3.8.*
